workshops/CVPR2025/pyproject.toml
[project]
name = "demo-cvpr-tutorial"
version = "0.1.0"
description = "Agentic Multimodal RAG Application"
readme = "README.md"
requires-python = ">=3.12"
keywords = [
    "multimodal",
    "rag", 
    "openvino",
    "video-search",
    "mcp",
    "agents",
    "ai",
    "llm",
    "vlm",
    "windows"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.12.1",
    "ipywidgets>=8.1.7",
    "llama-cloud>=0.1.32",
    "llama-cloud-services>=0.6.43",
    "llama-index>=0.12.51",
    "llama-index-agent-openai>=0.4.12",
    "llama-index-cli>=0.4.4",
    "llama-index-core>=0.12.52",
    "llama-index-embeddings-huggingface>=0.4.0",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-embeddings-openvino>=0.5.2",
    "llama-index-indices-managed-llama-cloud>=0.7.10",
    "llama-index-instrumentation>=0.3.0",
    "llama-index-llms-huggingface>=0.4.2",
    "llama-index-llms-openai>=0.4.7",
    "llama-index-llms-openvino>=0.4.0",
    "llama-index-multi-modal-llms-openai>=0.5.3",
    "llama-index-multi-modal-llms-openvino>=0.3.2",
    "llama-index-program-openai>=0.3.2",
    "llama-index-question-gen-openai>=0.3.1",
    "llama-index-readers-file>=0.4.11",
    "llama-index-readers-llama-parse>=0.4.0",
    "llama-index-vector-stores-qdrant>=0.6.1",
    "llama-index-workflows>=1.1.0",
    "llama-parse>=0.6.43",
    "llama-index-tools-mcp>=0.2.6",
    "transformers>=4.51.3",
    "moviepy>=2.2.1",
    "librosa>=0.11.0",
    "python-ffmpeg>=1.0.16",
    "open-clip-torch>=2.32.0",
    "huggingface-hub>=0.33.4",
    "gradio>=5.38.0",
    "opencv-python>=4.12.0.88",
    "webvtt-py>=0.5.1",
    "openvino-genai>=2025.2.0.0",
    "ffmpeg>=1.4",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "torch>=2.7.1",
    "lancedb>=0.24.1",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
